Reuse autocompressing inputstream in VersionsResource
